IOS-3014 Remove SupportChatService
