+++ Tangem/App/Models/UserWallet/SupportChatEnvironment.swift
@@ -1,0 +1,12 @@
+//
+//  SupportChatEnvironment.swift
+//  Tangem
+//
+//  Created by Pavel Grechikhin on 15.06.2022.
+//  Copyright © 2022 Tangem AG. All rights reserved.
+//
+
+enum SupportChatEnvironment {
+    case tangem
+    case saltPay
+}
