--- Tangem/App/Services/SupportChatService/SupportChatService.swift
@@ -1,51 +1,0 @@
-//
-//  ChatViewModel.swift
-//  Tangem
-//
-//  Created by Pavel Grechikhin on 15.06.2022.
-//  Copyright © 2022 Tangem AG. All rights reserved.
-//
-
-import Foundation
-import ZendeskCoreSDK
-import SupportSDK
-import MessagingSDK
-import ChatSDK
-import ChatProvidersSDK
-
-protocol SupportChatServiceProtocol {
-    func initialize(with env: SupportChatEnvironment)
-}
-
-// TODO: Move it initialize into SupportChatViewModel
-class SupportChatService: SupportChatServiceProtocol {
-    @Injected(\.keysManager) private var keysManager: KeysManager
-
-    func initialize(with env: SupportChatEnvironment) {
-        let config = keysManager.zendesk
-        Zendesk.initialize(
-            appId: config.zendeskAppId,
-            clientId: config.zendeskClientId,
-            zendeskUrl: config.zendeskUrl
-        )
-        Support.initialize(withZendesk: Zendesk.instance)
-        Zendesk.instance?.setIdentity(Identity.createAnonymous())
-        Chat.initialize(accountKey: config.zendeskAccountKey, appId: config.zendeskAppId)
-    }
-}
-
-enum SupportChatEnvironment {
-    case tangem
-    case saltPay
-}
-
-private struct KeysManagerKey: InjectionKey {
-    static var currentValue: SupportChatServiceProtocol = SupportChatService()
-}
-
-extension InjectedValues {
-    var supportChatService: SupportChatServiceProtocol {
-        get { Self[KeysManagerKey.self] }
-        set { Self[KeysManagerKey.self] = newValue }
-    }
-}
